Fix aggregation
