+++ geocoder/src/main/resources/application.conf
@@ -23,7 +23,7 @@ grasshopper {
     }
     metrics {
       delay = 2
-      interval = 5000
+      interval = 200
     }
   }
